+++ app/Http/Controllers/BahanBakarController.php
@@ -9,6 +9,7 @@ class BahanBakarController extends Controller
 {
     public function index()
     {
+        if ($redirect = $this->checkifLoginForStaff()) return $redirect;
         $bahanBakars = BahanBakar::latest()->paginate(5);
         $dataType = 'bahanBakar';
         // $bahanBakars = PerjalananKaryawanPerusahaan::all();
@@ -19,11 +20,13 @@ class BahanBakarController extends Controller
 
     public function add()
     {
+        if ($redirect = $this->checkifLoginForStaff()) return $redirect;
         return view('dashboardStaff.layouts.bahanBakar.add');
     }
 
     public function store(Request $request)
     {
+        if ($redirect = $this->checkifLoginForStaff()) return $redirect;
         $validatedData = $request->validate([
             'fuel_name' => 'required',
             'emission' => 'required',
@@ -42,12 +45,14 @@ class BahanBakarController extends Controller
 
     public function delete($id)
     {
+        if ($redirect = $this->checkifLoginForStaff()) return $redirect;
         BahanBakar::destroy($id);
         return redirect('dashboard/staff/bahanBakar')->with('success', 'Data Successfully Deleted');
     }
 
     public function edit($id)
     {
+        if ($redirect = $this->checkifLoginForStaff()) return $redirect;
         $oldData = BahanBakar::find($id);
 
         // return ($oldData);
@@ -57,6 +62,7 @@ class BahanBakarController extends Controller
 
     public function update(Request $request, string $id)
     {
+        if ($redirect = $this->checkifLoginForStaff()) return $redirect;
         $validatedData = $request->validate([
             'fuel_name' => 'required',
             'emission' => 'required',
@@ -75,6 +81,7 @@ class BahanBakarController extends Controller
 
     public function restore(string $id)
     {
+        if ($redirect = $this->checkifLoginForStaff()) return $redirect;
         BahanBakar::withTrashed()->where('id', $id)->restore();
         return redirect('dashboard/staff/bahanBakar')->with('success', 'Data Successfully Restored');
     }
